+++ projects/ngx-amap/src/types/class/riding/ride-route.ts
@@ -1,0 +1,7 @@
+import { RideStep } from './ride-step';
+
+export interface RideRoute {
+  distance: number;
+  time: number;
+  steps: RideStep[];
+}